Split components into submodule
