+++ src/oemof/solph/components/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+
+from .extraction_turbine_chp import ExtractionTurbineCHP  # noqa: F401
+from .generic_chp import GenericCHP  # noqa: F401
+from .generic_storage import GenericStorage  # noqa: F401
+from .offset_transformer import OffsetTransformer  # noqa: F401
